moving tests to their own package
